+++ frontend/Dockerfile
@@ -9,4 +9,4 @@ RUN npm run build
 
 FROM nginx:latest
 
-COPY --from=builder /app/dist /usr/share/nginx/html/
+COPY --from=builder /app/build /usr/share/nginx/html/
