+++ frontend/README.md
@@ -1,25 +1,38 @@
-# The frontend for ELEPHANT
+# create-svelte
 
-A simple frontend for elephant created with svelte, ts and vite
+Everything you need to build a Svelte project, powered by [`create-svelte`](https://github.com/sveltejs/kit/tree/master/packages/create-svelte).
 
-# Framework:
+## Creating a project
 
-Svelte + TS + Vite
+If you're seeing this, you've probably already done this step. Congrats!
 
-## Recommended IDE Setup
+```bash
+# create a new project in the current directory
+npm create svelte@latest
 
-[VS Code](https://code.visualstudio.com/) + [Svelte](https://marketplace.visualstudio.com/items?itemName=svelte.svelte-vscode).
+# create a new project in my-app
+npm create svelte@latest my-app
+```
 
-## Run and build
+## Developing
 
-### Install
+Once you've created a project and installed dependencies with `npm install` (or `pnpm install` or `yarn`), start a development server:
 
-`npm install`
+```bash
+npm run dev
 
-### Run
+# or start the server and open the app in a new browser tab
+npm run dev -- --open
+```
 
-`npm run dev`
+## Building
 
-### Build
+To create a production version of your app:
 
-`npm run build`
+```bash
+npm run build
+```
+
+You can preview the production build with `npm run preview`.
+
+> To deploy your app, you may need to install an [adapter](https://kit.svelte.dev/docs/adapters) for your target environment.
