--- frontend/environment/env.ts
@@ -1,4 +1,0 @@
-export const BACKEND = "http://localhost:3000";
-export const CORS_PROXY_ROUTE = "cors-proxy";
-export const MATCH_COUNT_HIGHLIGHT = 1500;
-export const CALENDAR_ROUTE = "calendar";
