--- frontend/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="de">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ELEPHANT</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
-  </body>
-</html>
