+++ frontend/src/lib/@components/index.ts
@@ -1,0 +1,1 @@
+// place files you want to import through the `$lib` alias in this folder.
