+++ frontend/src/routes/+layout.ts
@@ -1,0 +1,2 @@
+import '../app.css';
+export const prerender = true;
